feature: custom 404 page + rename context folder
