+++ interface/pages/NotFound.tsx
@@ -1,0 +1,9 @@
+export default function Custom404() {
+    return (
+        <div style={{zIndex: "-10", position: "absolute", display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center",width: "100vw", height: "100vh"}}>
+            <h2 style={{fontSize: "6rem"}}>404</h2>
+            <h3 style={{fontSize: "1.5rem", margin: "1rem 0"}}>🇫🇷 Page innexistante</h3>
+            <h3 style={{fontSize: "1.5rem", margin: "1rem 0"}}>🇺🇸 / 🇬🇧 Page not found</h3>
+        </div>
+    )
+}
